refactor(console): address CodeRabbit review on community packages + headers
- addonCategoryMeta: import ComponentType from react (no React namespace in a .ts file)
- AddonCard: extract StatChip into its own AddonStatChip.tsx (one component per file)
- AddonDetailDrawer: reset tab to overview when the addon changes; prefer fetched
  pkg metadata (displayName/description/tags/icon) over the summary addon
- PackagesPage: sanitize deriveNamespace to a valid [a-z0-9-]+ RPC namespace
- ApisPage/CredentialsPage: validate the tab URL param instead of force-casting
- ApisPage: translate the API-type switch labels via t() instead of asI18n literals
- FunctionDetailsForm: add aria-label to the icon-only edit button
- ShellHeader: split FilterChip/CycleSwitch/ActionCluster into their own files and
  move shared types/constants/helpers to shellHeaderShared; ShellHeader is now a
  const-arrow component, public types re-exported unchanged
